bbsvx_erlog_db_differ behaviour (bbsvx v0.1.0)

BBSvx Erlog Database Differ

Erlog database module proxy that logs all database modification operations.

Acts as a proxy for another erlog db module, tracking changes for transaction diffs.

Types

db()

-type db() :: #db{mod :: term(), ref :: term(), loc :: term()}.

erlog_state()

-opaque erlog_state()

functor()

-type functor() :: tuple().

Callbacks

abolish_clauses/2

-callback abolish_clauses(term(), functor()) -> error | {ok, term()}.

get_procedure(Db, Functor) -> built_in | {code,{Mod,Func}} | {clauses,[Clause]} | undefined. Return the procedure type and data for a functor.

add_built_in/2

-callback add_built_in(term(), functor()) -> term().

add_compiled_code(Db, Functor, Module, Function) -> {ok,Db} | error. Add functor as a compiled procedure with code in M:F in the database. Check that it is not a built-in, if so return error.

add_compiled_proc/4

-callback add_compiled_proc(term(), functor(), atom(), atom()) -> {ok, term()} | error.

asserta_clause(Db, Functor, Head, Body) -> {ok,NewDb} | error. assertz_clause(Db, Functor, Head, Body) -> {ok,NewDb} | error. Insert a clause at the beginning or end of the database.

asserta_clause/4

-callback asserta_clause(term(), functor(), term(), term()) -> {ok, term()} | error.

assertz_clause/4

-callback assertz_clause(term(), functor(), term(), term()) -> {ok, term()} | error.

retract_clause(Db, Functor, ClauseTag) -> {ok,NewDb} | error. Retract (remove) the clause with tag ClauseTag from the list of clauses of Functor.

get_interpreted_functors/1

-callback get_interpreted_functors(term()) -> [functor()].

get_procedure/2

-callback get_procedure(term(), functor()) ->
                           built_in | {code, {atom(), atom()}} | {clauses, [term()]} | undefined.

get_procedure_type(Db, Functor) -> built_in | compiled | interpreted | undefined. Return the procedure type for a functor.

get_procedure_type/2

-callback get_procedure_type(term(), functor()) -> built_in | compiled | interpreted | undefined.

get_interp_functors(Database) -> [Functor].

new(InitArgs)

-callback new(InitArgs :: term()) -> term().

add_built_in(Db, Functor) -> Db. Add functor as a built-in in the database.

retract_clause/3

-callback retract_clause(term(), functor(), term()) -> {ok, term()} | error.

Functions

abolish_clauses/2

add_built_in/2

add_compiled_proc/4

apply_diff/2

-spec apply_diff([term()], db()) -> {ok, db()}.

asserta_clause/4

assertz_clause/4

get_interpreted_functors/1

get_procedure/2

get_procedure_type/2

new(InitArgs)

retract_clause/3