bbsvx_ont_service (bbsvx v0.1.0)

BBSvx Ontology Service

Gen Server for managing ontology instances and Prolog knowledge base operations.

Provides API for ontology creation, querying, goal storage, and distributed connections.

Types

erlog_state()

-opaque erlog_state()

goal()

-type goal() ::
          #goal{id :: binary(),
                timestamp :: integer(),
                namespace :: binary(),
                source_id :: binary(),
                diff :: term(),
                payload :: term()}.

inet_address()

-type inet_address() :: inet:hostname() | inet:ip_address().

node_entry()

-type node_entry() ::
          #node_entry{node_id :: binary() | undefined,
                      host :: inet_address() | binary() | local,
                      port :: inet:port_number()}.

ontology()

-type ontology() ::
          #ontology{namespace :: binary(),
                    type :: shared | local,
                    version :: binary(),
                    last_update :: integer(),
                    contact_nodes :: [node_entry()]}.

Functions

binary_to_table_name(Namespace)

-spec binary_to_table_name(binary()) -> atom().

code_change(OldVsn, State, Extra)

connect_ontology(Namespace)

-spec connect_ontology(Namespace :: binary()) -> ok | {error, atom()}.

delete_ontology(Namespace)

-spec delete_ontology(Namespace :: binary()) -> ok | {error, atom()}.

disconnect_ontology(Namespace)

-spec disconnect_ontology(Namespace :: binary()) -> ok | {error, atom()}.

get_goal(GoalId)

get_ontology(Namespace)

-spec get_ontology(binary()) -> {ok, ontology()} | {error, atom()}.

handle_call/3

handle_cast(Msg, State)

handle_info(Info, State)

init/1

is_contact_node(NodeId, ContactNodes)

new_ontology/1

-spec new_ontology(ontology()) -> ok | {error, atom()}.

prove(Namespace, Predicate)

start_link()

-spec start_link() -> gen_statem:start_ret().

store_goal(Goal)

-spec store_goal(goal()) -> ok | {error, atom()}.

table_exists(Namespace)

terminate(Reason, State)